improvement: consolidated style files
