+++ web-client/src/modules/landing-page/components/figma.tsx
@@ -17,7 +17,7 @@ const Flex = styled.div`
 
 const Button = styled.button`
   height: 35px;
-  width: 130px;
+  width: 105px;
   border-radius: 4px;
   margin: 10px;
   cursor: pointer;
@@ -26,6 +26,18 @@ const Button = styled.button`
 // ---------------------------------------------------------------------------------------------------- //
 
 // Principal Components
+
+export const Nav = styled.nav`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  height: 65px;
+  position: fixed;
+  top: 0;
+  background: white;
+  z-index: 5;
+`;
 
 export const Root = styled.div`
   width: 100%;
@@ -56,8 +68,11 @@ export const MiddleSection = styled(Flex)`
 `;
 
 export const AttributesSection = styled(Container)``;
+
 export const HelpInfoSection = styled(Container)``;
+
 export const SponsorSection = styled(Container)``;
+
 export const VolunteerInfoSection = styled(Container)``;
 export const AboutInfoSection = styled(Container)``;
 export const SocialLinksSection = styled(Flex)``;
@@ -66,6 +81,21 @@ export const LegalSection = styled(Flex)``;
 // ---------------------------------------------------------------------------------------------------- //
 
 // Containers
+
+export const LogoContainer = styled.div`
+  margin-left: 3vw;
+`;
+
+export const ButtonContainer = styled.div`
+  margin-right: 3vw;
+`;
+
+export const LinkContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  flex: 1;
+`;
 
 export const TextTopContainer = styled(Container)`
   position: relative;
@@ -84,6 +114,49 @@ export const AboutTextContainer = styled(Container)``;
 // ---------------------------------------------------------------------------------------------------- //
 
 // Child Components
+
+export const Link = styled.a`
+  text-decoration: none;
+  margin: 0px 15px;
+  color: black;
+`;
+
+export const Dropdown = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding-left: 5px;
+  padding-right: 0;
+
+  ul {
+    list-style: none;
+    padding-left: 2px;
+    height: 25px;
+    overflow: hidden;
+    position: relative;
+    top: 5px;
+    width: 80px;
+    background: white;
+    border-radius: 5px;
+
+    &: hover {
+      position: relative;
+      top: 40px;
+      height: 95px;
+    }
+
+    li {
+      cursor: pointer;
+      border-radius: 5px;
+      padding: 4px;
+      margin-right: 2px;
+      &: hover {
+        color: white;
+        background: #ffcb52;
+      }
+    }
+  }
+`;
 
 export const TextTop = styled.div`
   h1 {
@@ -132,6 +205,23 @@ export const VolunteerInfoCard = styled(Container)``;
 
 // Buttons
 
+export const LogInButton = styled(Button)`
+  background: none;
+  border: none;
+  color: #eb7100;
+  transition: all 0.3s ease;
+
+  &:hover {
+    border: 1px solid #eb7100;
+  }
+`;
+
+export const SignUpButton = styled(Button)`
+  border: none;
+  color: white;
+  background: #ff7b02;
+`;
+
 export const GetHelpBtn = styled(Button)`
   color: white;
   background: #ff7b02;
@@ -153,6 +243,17 @@ export const VolunteerBtn = styled(Button)`
 // ---------------------------------------------------------------------------------------------------- //
 
 // Image Components
+
+export const Logo = styled.img``;
+
+export const BetaFlair = styled.img`
+  height: 20px;
+  border-radius: 5px;
+  margin-left: 10px;
+  position: relative;
+  bottom: 12px;
+  font-size: 14px;
+`;
 
 export const RectTop = styled.img`
   position: absolute;
